refactor(viewer): make naming consistent with Typescript standards + remove dead or duplicate code
