+++ viewer/src/app/feature-view/fullBoxControl.ts
@@ -1,17 +1,17 @@
 import { Control } from 'ol/control.js'
 
 import { EventEmitter } from '@angular/core'
-import { emitBox } from './boxcontrol'
+import { emitBox } from './boxControl'
 import { Geometry } from 'ol/geom'
 import { fromExtent } from 'ol/geom/Polygon'
 
-export class fullBoxControl extends Control {
+export class FullBoxControl extends Control {
   constructor(
     public boxEmitter: EventEmitter<string>,
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    opt_options: any
+    optionalOptions: any
   ) {
-    const options = opt_options || {}
+    const options = optionalOptions || {}
 
     const button = document.createElement('button')
