+++ viewer/src/app/object-info/object-info.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, ViewEncapsulation } from '@angular/core'
+import { ChangeDetectionStrategy, Component, Input, ViewEncapsulation } from '@angular/core'
 import { CommonModule } from '@angular/common'
 import RenderFeature from 'ol/render/Feature'
 
@@ -12,6 +12,7 @@ type propRow = {
   standalone: true,
   encapsulation: ViewEncapsulation.ShadowDom,
   imports: [CommonModule],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   templateUrl: './object-info.component.html',
   styleUrls: ['./object-info.component.css'],
 })
